backend/requirements.txt
fastapi
uvicorn
pdfminer.six
pymupdf
sentence-transformers
scikit-learn
google-generativeai
openai
python-multipart
